change: split UniversimeApi into different modules
