+++ src/contexts/Auth/AuthProvider.tsx
@@ -41,7 +41,7 @@ export const AuthProvider = ({ children }: { children: JSX.Element }) => {
   }
 
   async function signin(email: string, password: string) {
-      const data = await UniversimeApi.signin(email, password);
+      const data = await UniversimeApi.Auth.signin(email, password);
       const profile = await UniversimeApi.Profile.profile();
       return setLoggedUser(data.body.user, data.token, profile.body.profile);
   }
