+++ src/services/UniversimeApi/index.ts
@@ -1,4 +1,17 @@
-import { UniversimeApi } from "./UniversimeApi";
+import * as Auth from "./Auth";
+import * as Competence from "./Competence";
+import * as CompetenceType from "./CompetenceType";
+import * as Group from "./Group";
+import * as Link from "./Link";
+import * as Profile from "./Profile";
+
+export const UniversimeApi = {
+    Auth,
+    Competence,
+    CompetenceType,
+    Group,
+    Link,
+    Profile,
+};
 
 export default UniversimeApi;
-export * from "./UniversimeApi";
